refactor(automations): extract schedule list
